chore(ts): export types properly
